Split the pty keepalive table
procemu.c had grown to 4759 lines carrying three interceptors and a pty
side table that shares nothing with them: /proc, /sys and /dev synthesis
on one side, and on the other a keepalive table, a shared segment, slave
accounting and the /dev/pts staging directory.

The pty half moves out whole. procemu-internal.h carries the eight
declarations the two halves still exchange, and the scratch-directory
registry gains a named entry point so the pty side reaches it without
the five statics behind it.
